déplace des fonctions utilitaires pour les exports
